Verify secrets are writable before we generate keys.
Otherwise we end up in a retry loop where we keep cycling through kvnos.
